+++ internal/command/interactive/cursor.go
@@ -1,0 +1,20 @@
+package interactive
+
+type cursor uint16
+
+func (c *cursor) up() {
+	if *c > 0 {
+		*c--
+	}
+}
+
+func (c *cursor) down(threshold int) {
+	if *c < cursor(threshold) {
+		*c++
+	}
+}
+
+func (c *cursor) setStart() {
+	var zero cursor
+	c = &zero
+}
